refactor: jsx 컴포넌트 작업 완료
